+++ packages/solana-wallet-snap/src/core/handlers/onKeyringRequest/Keyring.test.ts
@@ -103,7 +103,8 @@ describe('SolanaKeyring', () => {
     mockAssetsService = {
       fetch: jest.fn().mockResolvedValue(MOCK_ASSET_ENTITIES),
       saveMany: jest.fn(),
-      findByAccount: jest.fn(),
+      getAccountAssets: jest.fn(),
+      getAccountAssetsByIDs: jest.fn(),
       getNativeAssetTypes: jest
         .fn()
         .mockReturnValue([KnownCaip19Id.SolMainnet]),
@@ -143,13 +144,16 @@ describe('SolanaKeyring', () => {
   describe('getAccountAssets', () => {
     it('calls the assets service', async () => {
       jest
-        .spyOn(mockAssetsService, 'findByAccount')
+        .spyOn(mockAssetsService, 'getAccountAssets')
         .mockResolvedValue(MOCK_ASSET_ENTITIES);
 
       const result = await keyring.getAccountAssets(
         MOCK_SOLANA_KEYRING_ACCOUNT_0.id,
       );
 
+      expect(mockAssetsService.getAccountAssets).toHaveBeenCalledWith(
+        MOCK_SOLANA_KEYRING_ACCOUNT_0.id,
+      );
       expect(result).toStrictEqual([
         MOCK_ASSET_ENTITY_0.assetType,
         MOCK_ASSET_ENTITY_1.assetType,
@@ -158,7 +162,7 @@ describe('SolanaKeyring', () => {
     });
 
     it('removes token assets with zero balance', async () => {
-      jest.spyOn(mockAssetsService, 'findByAccount').mockResolvedValue([
+      jest.spyOn(mockAssetsService, 'getAccountAssets').mockResolvedValue([
         MOCK_ASSET_ENTITY_1, // Token asset with non-zero balance
         { ...MOCK_ASSET_ENTITY_2, rawAmount: '0' }, // Token asset with zero balance
       ]);
@@ -171,7 +175,7 @@ describe('SolanaKeyring', () => {
     });
 
     it('keeps the native asset even if it has zero balance', async () => {
-      jest.spyOn(mockAssetsService, 'findByAccount').mockResolvedValue([
+      jest.spyOn(mockAssetsService, 'getAccountAssets').mockResolvedValue([
         { ...MOCK_ASSET_ENTITY_0, rawAmount: '0' }, // Native asset with zero balance
         { ...MOCK_ASSET_ENTITY_1, rawAmount: '0' }, // Token asset with zero balance
       ]);
@@ -343,9 +347,9 @@ describe('SolanaKeyring', () => {
         symbol: 4,
       } as unknown as AssetEntity;
 
-      jest
-        .spyOn(mockAssetsService, 'findByAccount')
-        .mockResolvedValue([invalidAsset]);
+      jest.spyOn(mockAssetsService, 'getAccountAssetsByIDs').mockResolvedValue({
+        [KnownCaip19Id.SolMainnet]: invalidAsset,
+      });
 
       await expect(
         keyring.getAccountBalances(MOCK_SOLANA_KEYRING_ACCOUNT_1.id, [
@@ -355,10 +359,13 @@ describe('SolanaKeyring', () => {
     });
 
     it('removes token assets with zero balance', async () => {
-      jest.spyOn(mockAssetsService, 'findByAccount').mockResolvedValue([
-        MOCK_ASSET_ENTITY_1, // Token asset with non-zero balance
-        { ...MOCK_ASSET_ENTITY_2, rawAmount: '0' }, // Token asset with zero balance
-      ]);
+      jest.spyOn(mockAssetsService, 'getAccountAssetsByIDs').mockResolvedValue({
+        [MOCK_ASSET_ENTITY_1.assetType]: MOCK_ASSET_ENTITY_1,
+        [MOCK_ASSET_ENTITY_2.assetType]: {
+          ...MOCK_ASSET_ENTITY_2,
+          rawAmount: '0',
+        },
+      });
 
       const result = await keyring.getAccountBalances(
         MOCK_SOLANA_KEYRING_ACCOUNT_0.id,
@@ -374,10 +381,16 @@ describe('SolanaKeyring', () => {
     });
 
     it('keeps the native asset even if it has zero balance', async () => {
-      jest.spyOn(mockAssetsService, 'findByAccount').mockResolvedValue([
-        { ...MOCK_ASSET_ENTITY_0, rawAmount: '0' }, // Native asset with zero balance
-        { ...MOCK_ASSET_ENTITY_1, rawAmount: '0' }, // Token asset with zero balance
-      ]);
+      jest.spyOn(mockAssetsService, 'getAccountAssetsByIDs').mockResolvedValue({
+        [MOCK_ASSET_ENTITY_0.assetType]: {
+          ...MOCK_ASSET_ENTITY_0,
+          rawAmount: '0',
+        },
+        [MOCK_ASSET_ENTITY_1.assetType]: {
+          ...MOCK_ASSET_ENTITY_1,
+          rawAmount: '0',
+        },
+      });
 
       const result = await keyring.getAccountBalances(
         MOCK_SOLANA_KEYRING_ACCOUNT_0.id,